work on routes, controllers, and dashboard
