+++ public/css/styles.css
@@ -8692,7 +8692,6 @@ readers do not read off random characters that represent icons */
 }
 
 #dashboard {
-  display: flex;
   font-family: "Nunito", sans-serif;
 }
 #dashboard a {
@@ -8708,7 +8707,7 @@ readers do not read off random characters that represent icons */
 #dashboard .main {
   min-height: 100vh;
   width: 100%;
-  overflow: hidden;
+  overflow: scroll;
   transition: all 0.35s ease-in-out;
   background-color: #fafbfe;
 }
